+++ meson.build
@@ -214,6 +214,7 @@ sources = [
   'src/game/objects/general/trapdoor.c',
   'src/game/objects/general/waterfall.c',
   'src/game/objects/names.c',
+  'src/game/objects/setup.c',
   'src/game/objects/traps/damocles_sword.c',
   'src/game/objects/traps/dart.c',
   'src/game/objects/traps/dart_emitter.c',
@@ -258,7 +259,6 @@ sources = [
   'src/game/savegame/savegame_bson.c',
   'src/game/savegame/savegame_legacy.c',
   'src/game/screen.c',
-  'src/game/setup.c',
   'src/game/shell.c',
   'src/game/sound.c',
   'src/game/stats.c',
