+++ src/game/objects/setup.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "global/types.h"
+
+void Object_SetupAllObjects(void);
