--- src/game/setup.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "global/types.h"
-
-void Setup_Creatures(void);
-void Setup_Traps(void);
-void Setup_MiscObjects(void);
-void Setup_AllObjects(void);
